Refactor ExportUtil into model.export package
