+++ src/main/java/seedu/address/logic/commands/ExportCommand.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.Messages;
-import seedu.address.commons.util.ExportUtil;
+import seedu.address.model.export.ExportUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.category.Category;
